+++ internal/router/connection.go
@@ -1,4 +1,4 @@
-package database
+package router
 
 import (
 	"google.golang.org/grpc"
@@ -7,11 +7,11 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/closer"
 )
 
-type Cluster interface {
-	// ClientConnInterface interface allows Cluster use as grpc.ClientConnInterface
+type Router interface {
+	// ClientConnInterface interface allows Router use as grpc.ClientConnInterface
 	// with lazy getting raw grpc-connection in Invoke() or NewStream() stages.
-	// Lazy getting grpc-connection must use for embedded client-side balancing
-	// DB may be put into code-generated client constructor as is.
+	// Lazy getting grpc-connection must use for embedded into driver client-side
+	// balancing may be put into code-generated client constructor as is.
 	grpc.ClientConnInterface
 	closer.Closer
 }
@@ -24,15 +24,15 @@ type Info interface {
 	// Endpoint returns initial endpoint
 	Endpoint() string
 
-	// Name returns database name
+	// Name returns router name
 	Name() string
 
-	// Secure returns true if database connection is secure
+	// Secure returns true if router connection is secure
 	Secure() bool
 }
 
 type Connection interface {
-	Cluster
 	Info
+	Router
 	Discoverer
 }