+++ internal/table/client_test.go
@@ -35,7 +35,7 @@ func TestSessionPoolCreateAbnormalResult(t *testing.T) {
 	defer cancel()
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -93,7 +93,7 @@ func TestSessionPoolKeeperWake(t *testing.T) {
 
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -169,7 +169,7 @@ func TestSessionPoolCloseWhenWaiting(t *testing.T) {
 			)
 			p := newClientWithStubBuilder(
 				t,
-				testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+				testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 					// nolint:unparam
 					testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 						return &Ydb_Table.CreateSessionResult{
@@ -247,7 +247,7 @@ func TestSessionPoolClose(t *testing.T) {
 	wg := sync.WaitGroup{}
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+		testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 			// nolint:unparam
 			testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 				return &Ydb_Table.CreateSessionResult{
@@ -344,7 +344,7 @@ func TestRaceWgClosed(t *testing.T) {
 			wg := sync.WaitGroup{}
 			p := newClientWithStubBuilder(
 				t,
-				testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+				testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 					// nolint:unparam
 					testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 						return &Ydb_Table.CreateSessionResult{
@@ -401,7 +401,7 @@ func TestSessionPoolDeleteReleaseWait(t *testing.T) {
 			)
 			p := newClientWithStubBuilder(
 				t,
-				testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+				testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 					// nolint:unparam
 					testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 						return &Ydb_Table.CreateSessionResult{
@@ -545,7 +545,7 @@ func TestSessionPoolRacyGet(t *testing.T) {
 func TestSessionPoolPutInFull(t *testing.T) {
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+		testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 			// nolint:unparam
 			testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 				return &Ydb_Table.CreateSessionResult{
@@ -593,7 +593,7 @@ func TestSessionPoolSizeLimitOverflow(t *testing.T) {
 			)
 			p := newClientWithStubBuilder(
 				t,
-				testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+				testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 					// nolint:unparam
 					testutil.TableCreateSession: func(interface{}) (result proto.Message, _ error) {
 						return &Ydb_Table.CreateSessionResult{
@@ -694,7 +694,7 @@ func TestSessionPoolGetPut(t *testing.T) {
 	}
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -741,7 +741,7 @@ func TestSessionPoolDisableBackgroundGoroutines(t *testing.T) {
 
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+		testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 			// nolint:unparam
 			testutil.TableCreateSession: func(interface{}) (result proto.Message, _ error) {
 				return &Ydb_Table.CreateSessionResult{
@@ -779,7 +779,7 @@ func TestSessionPoolKeepAlive(t *testing.T) {
 	)
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -856,7 +856,7 @@ func TestSessionPoolKeepAliveOrdering(t *testing.T) {
 	)
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -925,7 +925,7 @@ func TestSessionPoolKeepAliveOrdering(t *testing.T) {
 func TestSessionPoolDoublePut(t *testing.T) {
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
+		testutil.NewRouter(testutil.WithInvokeHandlers(testutil.InvokeHandlers{
 			// nolint:unparam
 			testutil.TableCreateSession: func(interface{}) (proto.Message, error) {
 				return &Ydb_Table.CreateSessionResult{
@@ -964,7 +964,7 @@ func TestSessionPoolKeepAliveCondFairness(t *testing.T) {
 	)
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -1048,7 +1048,7 @@ func TestSessionPoolKeepAliveMinSize(t *testing.T) {
 	idleThreshold := 5 * time.Second
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -1115,7 +1115,7 @@ func TestSessionPoolKeepAliveMinSize(t *testing.T) {
 func TestSessionPoolKeepAliveWithBadSession(t *testing.T) {
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -1161,7 +1161,7 @@ func TestSessionPoolKeeperRetry(t *testing.T) {
 	retry := true
 	p := newClientWithStubBuilder(
 		t,
-		testutil.NewDB(
+		testutil.NewRouter(
 			testutil.WithInvokeHandlers(
 				testutil.InvokeHandlers{
 					// nolint:unparam
@@ -1283,7 +1283,7 @@ var okHandler = func(interface{}) (proto.Message, error) {
 	return nil, nil
 }
 
-var simpleCluster = testutil.NewDB(
+var simpleCluster = testutil.NewRouter(
 	testutil.WithInvokeHandlers(
 		testutil.InvokeHandlers{
 			// nolint:unparam